removed the previous email handler code
