+++ app/templates/events/event_list.html
@@ -30,7 +30,6 @@
     </div>
   </div>
 </div>
-
 <div>
   <ul class="nav nav-tabs nav-fill mx-3 mb-3" id="pills-tab" role="tablist">
   <li class="nav-item" role="presentation">
@@ -51,7 +50,6 @@
   </li>
   </ul>
 </div>
-
 {% macro createTable(events) %}
 {% if events.exists() %}
   <table class="table table-striped">
@@ -65,7 +63,6 @@
       </tr>
     </thead>
     <tbody>
-
         {% for event in events %}
           <tr>
             <td> <a href="/event/{{event.id}}/edit" class="link-primary">{{event.name}}</a></td>
@@ -74,7 +71,6 @@
             <td>{{event.location}}</td>
             {% if user.isAdmin %}
               <td>
-
                   <button class="btn btn-primary" onclick="showEmailModal({{event.id}}, 'Unknown', {{selectedTerm}})"{{ 'disabled' if event.isPast }}>
                     <span class="bi bi-envelope-fill"> Email</span>
                   </button>
@@ -96,13 +92,11 @@
   <p>No upcoming events for this term</p>
 {% endif %}
 {% endmacro %}
-
 <div class="tab-content" id="pills-tabContent">
   <div class="tab-pane fade show active" id="pills-student-led" role="tabpanel" aria-labelledby="pills-student-led-tab">
     {% if studentLedProgram %}
     <div class="accordion" id="categoryAccordion">
       {% for program,events in studentLedProgram.items() %}
-
         <div class="accordion-item">
           <h2 class="accordion-header" id="accordion__header_{{program}}">
             <button class="accordion-button collapsed"
@@ -138,7 +132,6 @@
                         <td>{{event.location}}</td>
                         {% if user.isAdmin %}
                           <td>
-
                               <button class="btn btn-primary"
                               onclick="showEmailModal({{event.id}}, {{program.id}}, {{selectedTerm}})" {{ 'disabled' if event.isPast }}>
                                 <span class="bi bi-envelope-fill"> Email</span>
@@ -166,10 +159,8 @@
       <p>No upcoming events for this term</p>
     {% endif %}
   </div>
-
   <div class="tab-pane fade show" id="pills-training" role="tabpanel" aria-labelledby="pills-training-tab">
     {{createTable(trainingProgram)}}
-
   </div>
   <div class="tab-pane fade show" id="pills-bonner-scholars" role="tabpanel" aria-labelledby="pills-bonner-scholars-tab">
     {{createTable(bonnerProgram)}}
@@ -178,7 +169,5 @@
     {{createTable(oneTimeEvents)}}
   </div>
 </div>
-
 {% include 'emailModal.html' %}
-
 {% endblock %}
